Ensure config env vars read at startup
The technique used is a bit of a hack: at Application startup, right
before starting the supervision tree, read the necessary environment
variables (with defaults) and persist them in the Application config.
Roughly follows Sasa Juric's method described in
https://elixirforum.com/t/what-is-the-difference-between-using-system-port-and-system-get-env-port-in-deployment/1975/4
.

The alternative was to use one of the methods described in
https://hexdocs.pm/distillery/config/runtime.html , but none of them
felt quite right for my use case.

Also move out the response JSON schema validation function into the
`Response` module in preparation for more testing.